analyse/requirements.txt
psycopg2-binary
numpy
